move test out of storage to core execution tests
